requirements.txt
flask
flask-socketio
paho-mqtt
requests
opencv-python
matplotlib
